Add notes about breakcond required registers
The idea is that each code structure returns the registers needed to perform
that operation. If any 8 bit register can be used 'x' is returned, if any 16
bit register can be used 'xx' is returned. The caller can then eliminate the
specific registers that are needed (eg: 'a', 'hl') and be left with a count of
'xx' or 'x' registers required. For example if 'a' and 'hl' and 'xx' are
required then 'a' and 'hl' are marked as in use for that statement and one 16
bit register (either 'bc' or 'de') is also required. This can be used to
generate "paths" of free register counts along the list of statements. This may
mean a push/pop of an earlier variable may be avoided by just keeping that
value in a free register along the path.

The problem arises when 16 bit registers can also be used as 8 bit registers.
This means the number of free 8 bit registers is dependent on the number of
free 16 bit registers. Counting one free 16 bit register must also decrease the
count of free 8 bit registers by 2.